move runPrompt to global scope only
